[!!!][TASK] Use packagist's swiftmailer
This removes the hardcopied swiftmailer resource files
from the package and adds a required-dependency to
the composer.json in order to let composer fetch it.

Additionally, it raises the swiftmailer version in use
from version 4.1.5 to 4.1.*.

This is marked breaking in case of the resource files
having been reference manually somewhere.

Change-Id: Ife7f04807bf00fd6f6a06abbe21682829b8c2a87